+++ sample-apps/react-native/dogfood/env.d.ts
@@ -1,4 +1,0 @@
-declare module 'react-native-dotenv' {
-  export const STREAM_API_KEY: string;
-  export const STREAM_API_SECRET: string;
-}