+++ sample-apps/react-native/expo-video-sample/components/VideoWrapper.tsx
@@ -5,7 +5,7 @@ import {
   StreamVideoClient,
 } from '@stream-io/video-react-native-sdk';
 import { useAppContext } from '../context/AppContext';
-import { STREAM_API_KEY } from '../data/constants';
+import { createToken } from '../utils/createToken';
 
 export const VideoWrapper = ({ children }: PropsWithChildren<{}>) => {
   const { user } = useAppContext();
@@ -14,16 +14,22 @@ export const VideoWrapper = ({ children }: PropsWithChildren<{}>) => {
   );
 
   useEffect(() => {
-    const _videoClient = new StreamVideoClient({
-      apiKey: STREAM_API_KEY,
-      user,
-      tokenProvider: user?.custom?.token,
-      options: { logLevel: 'warn' },
-    });
-    setVideoClient(_videoClient);
+    let _videoClient: StreamVideoClient | undefined;
+    const run = async () => {
+      if (!user?.id) return;
+      const { token, apiKey } = await createToken({ user_id: user.id });
+      _videoClient = new StreamVideoClient({
+        apiKey,
+        user,
+        token,
+        options: { logLevel: 'warn' },
+      });
+      setVideoClient(_videoClient);
+    };
+    run();
 
     return () => {
-      _videoClient.disconnectUser();
+      _videoClient?.disconnectUser();
       setVideoClient(undefined);
     };
   }, [user]);