+++ sample-apps/react-native/expo-video-sample/data/users.ts
@@ -4,28 +4,16 @@ export const users = [
     name: 'Khushal Agarwal',
     imageUrl:
       'https://ca.slack-edge.com/T02RM6X6B-U02DTREQ2KX-5d600c87d3bc-512',
-    custom: {
-      token:
-        'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VyX2lkIjoia2h1c2hhbGV4cG8ifQ.2Ow87A66ZuxY7A68q65CIUtYSSU7Diwvx0p4SHtkvc0',
-    },
   },
   {
     id: 'santhoshexpo',
     name: 'Santhosh Vaiyapuri',
     imageUrl:
       'https://ca.slack-edge.com/T02RM6X6B-U0359AX2TUY-dc7dbec0bb88-512',
-    custom: {
-      token:
-        'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VyX2lkIjoic2FudGhvc2hleHBvIn0.b4MJm_7PemLXWlX7So-BgpeQ6w8cOcX_jZMYhXJGfY4',
-    },
   },
   {
     id: 'vishalexpo',
     name: 'Vishal Narkhede',
     imageUrl: 'https://ca.slack-edge.com/T02RM6X6B-UHGDQJ8A0-b4a6ca584e05-512',
-    custom: {
-      token:
-        'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VyX2lkIjoidmlzaGFsZXhwbyJ9.xN2X3gBoEtbQonZ9aIhWBAoeSXhYsySAVD8GXP-XADw',
-    },
   },
 ];